gestaltd: delete unused v1 REST endpoints migrated to SDK gRPC
Deletes 16 handwritten v1 REST handlers that the CLI has fully migrated
to SDK gRPC sync methods. The CLI no longer calls these endpoints —
they are dead code.

Agent CRUD (11 endpoints) → AgentClient sync methods:
- POST/GET /api/v1/agent/sessions, GET/PATCH /api/v1/agent/sessions/{id}
- POST/GET /api/v1/agent/sessions/{id}/turns
- GET /api/v1/agent/turns/{id}, POST /api/v1/agent/turns/{id}/cancel
- GET /api/v1/agent/turns/{id}/events
- GET /api/v1/agent/turns/{id}/interactions
- POST /api/v1/agent/turns/{id}/interactions/{id}/resolve

Workflow runs (3 endpoints) → WorkflowClient sync methods:
- GET /api/v1/workflow/runs, GET /api/v1/workflow/runs/{id}
- POST /api/v1/workflow/runs/{id}/cancel

Token list/revoke-all (2 endpoints) → IdentityClient sync methods:
- GET /api/v1/tokens (list)
- DELETE /api/v1/tokens (revoke all)

Keeping: POST /api/v1/tokens (create) and DELETE /api/v1/tokens/{id}
(revoke single) — the auth login/logout flow still uses them.
Keeping: GET /api/v1/agent/turns/{id}/events/stream (SSE),
GET /api/v1/agent/providers, POST /api/v1/agent/harnesses/resolve —
no proto RPC exists for these.

Deletes agent_session_turn_test.go (1943 lines, all tests set up via
deleted CRUD endpoints), workflow_run_test.go (370 lines), and 10 token
list/revoke-all test functions from server_test.go. Extracts shared test
helpers (stubAgentControl, memoryAgentProvider) into
agent_test_helpers_test.go.

13 files changed, 735 insertions, 4562 deletions.
