+++ gestaltd/internal/server/grant_helpers.go
@@ -19,14 +19,6 @@ type createGrantResponse struct {
 	ExpiresAt *time.Time `json:"expiresAt,omitempty"`
 }
 
-type grantInfo struct {
-	ID        string   `json:"id"`
-	Name      string   `json:"name,omitempty"`
-	Scopes    []string `json:"scopes,omitempty"`
-	CreatedAt string   `json:"createdAt,omitempty"`
-	ExpiresAt string   `json:"expiresAt,omitempty"`
-}
-
 func (s *Server) callerAuthContext(ctx context.Context, r *http.Request) context.Context {
 	token, err := requestSessionOrBearerToken(r)
 	if err != nil || strings.TrimSpace(token) == "" {
@@ -44,28 +36,6 @@ func (s *Server) callerBearerToken(r *http.Request) (string, error) {
 		return "", errors.New("caller bearer token required")
 	}
 	return strings.TrimSpace(token), nil
-}
-
-func grantUnixSecondsToRFC3339(seconds int64) string {
-	if seconds <= 0 {
-		return ""
-	}
-	return time.Unix(seconds, 0).UTC().Format(time.RFC3339)
-}
-
-func grantInfoFromResponse(grantID string, resp *core.GetGrantResponse) grantInfo {
-	info := grantInfo{ID: grantID, Name: grantID}
-	if resp == nil {
-		return info
-	}
-	info.CreatedAt = grantUnixSecondsToRFC3339(resp.CreatedAt)
-	info.ExpiresAt = grantUnixSecondsToRFC3339(resp.ExpiresAt)
-	for _, scope := range resp.Scopes {
-		if scope.Scope != "" {
-			info.Scopes = append(info.Scopes, scope.Scope)
-		}
-	}
-	return info
 }
 
 func tokenExpiresAt(now func() time.Time, expiresIn int) *time.Time {
