+++ gestaltd/internal/server/server_test.go
@@ -9397,94 +9397,6 @@ func TestIntegrationOAuthCallback_InvalidState(t *testing.T) {
 	})
 }
 
-func TestCreateAndListAPITokens(t *testing.T) {
-	t.Parallel()
-
-	ts := newTestServer(t, func(cfg *server.Config) {
-		configureGrantTestAuth(cfg)
-		cfg.Providers = grantTestProviders(t)
-		cfg.Services = testutil.NewStubServices(t)
-	})
-	testutil.CloseOnCleanup(t, ts)
-
-	body := bytes.NewBufferString(`{"name":"my-token","scopes":"testapp"}`)
-	req, _ := http.NewRequest(http.MethodPost, ts.URL+"/api/v1/tokens", body)
-	req.Header.Set("Content-Type", "application/json")
-	addGrantTestSessionCookie(req)
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		t.Fatalf("request: %v", err)
-	}
-	defer func() { _ = resp.Body.Close() }()
-
-	if resp.StatusCode != http.StatusCreated {
-		t.Fatalf("expected 201, got %d", resp.StatusCode)
-	}
-
-	var result struct {
-		ID     string   `json:"id"`
-		Token  string   `json:"token"`
-		Name   string   `json:"name"`
-		Scopes []string `json:"scopes"`
-	}
-	if err := json.NewDecoder(resp.Body).Decode(&result); err != nil {
-		t.Fatalf("decoding: %v", err)
-	}
-	if result.Token == "" {
-		t.Fatal("expected non-empty token in response")
-	}
-	if result.ID == "" {
-		t.Fatal("expected non-empty grant id in response")
-	}
-	if result.Name != result.ID {
-		t.Fatalf("expected name to equal grant id %q, got %q", result.ID, result.Name)
-	}
-	if len(result.Scopes) != 1 || result.Scopes[0] != "testapp" {
-		t.Fatalf("expected scopes [testapp], got %v", result.Scopes)
-	}
-
-	listReq, _ := http.NewRequest(http.MethodGet, ts.URL+"/api/v1/tokens", nil)
-	addGrantTestSessionCookie(listReq)
-	listResp, err := http.DefaultClient.Do(listReq)
-	if err != nil {
-		t.Fatalf("list request: %v", err)
-	}
-	defer func() { _ = listResp.Body.Close() }()
-	if listResp.StatusCode != http.StatusOK {
-		respBody, _ := io.ReadAll(listResp.Body)
-		t.Fatalf("list status = %d, want 200: %s", listResp.StatusCode, respBody)
-	}
-
-	var listed []struct {
-		ID        string   `json:"id"`
-		Name      string   `json:"name"`
-		Scopes    []string `json:"scopes"`
-		CreatedAt string   `json:"createdAt"`
-		ExpiresAt string   `json:"expiresAt"`
-	}
-	if err := json.NewDecoder(listResp.Body).Decode(&listed); err != nil {
-		t.Fatalf("decode list response: %v", err)
-	}
-	if len(listed) != 1 {
-		t.Fatalf("expected 1 listed token, got %d", len(listed))
-	}
-	if listed[0].ID != result.ID {
-		t.Fatalf("listed id = %q, want %q", listed[0].ID, result.ID)
-	}
-	if listed[0].Name != result.ID {
-		t.Fatalf("listed name = %q, want grant id %q", listed[0].Name, result.ID)
-	}
-	if len(listed[0].Scopes) != 1 || listed[0].Scopes[0] != "testapp" {
-		t.Fatalf("listed scopes = %v, want [testapp]", listed[0].Scopes)
-	}
-	if _, err := time.Parse(time.RFC3339, listed[0].CreatedAt); err != nil {
-		t.Fatalf("listed createdAt = %q, want RFC3339: %v", listed[0].CreatedAt, err)
-	}
-	if _, err := time.Parse(time.RFC3339, listed[0].ExpiresAt); err != nil {
-		t.Fatalf("listed expiresAt = %q, want RFC3339: %v", listed[0].ExpiresAt, err)
-	}
-}
-
 func TestCreateAPITokenAllowsEmptyScopes(t *testing.T) {
 	t.Parallel()
 
@@ -9605,110 +9517,6 @@ func TestCreateAPITokenDoesNotUseStateForName(t *testing.T) {
 	}
 }
 
-func TestListAPITokensSkipsNotFoundGrant(t *testing.T) {
-	t.Parallel()
-
-	stub := newGrantTrackingAuthStub()
-	now := time.Now().UTC()
-	stub.grants["grant-live"] = &core.GetGrantResponse{
-		CreatedAt: now.Unix(),
-		ExpiresAt: now.Add(time.Hour).Unix(),
-		Scopes:    []core.GrantScope{{Scope: "testapp"}},
-	}
-	stub.ListGrantsFn = func(context.Context, *core.ListGrantsRequest) (*core.ListGrantsResponse, error) {
-		return &core.ListGrantsResponse{GrantIDs: []string{"grant-live", "grant-stale"}}, nil
-	}
-	stub.GetGrantFn = func(_ context.Context, req *core.GetGrantRequest) (*core.GetGrantResponse, error) {
-		if req == nil {
-			return nil, fmt.Errorf("missing grant request")
-		}
-		switch req.GrantID {
-		case "grant-live":
-			return stub.grants["grant-live"], nil
-		case "grant-stale":
-			return nil, core.ErrNotFound
-		default:
-			return nil, core.ErrNotFound
-		}
-	}
-
-	ts := newTestServer(t, func(cfg *server.Config) {
-		cfg.Auth = stub
-		cfg.Services = testutil.NewStubServices(t)
-	})
-	testutil.CloseOnCleanup(t, ts)
-
-	req, _ := http.NewRequest(http.MethodGet, ts.URL+"/api/v1/tokens", nil)
-	addGrantTestSessionCookie(req)
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		t.Fatalf("list request: %v", err)
-	}
-	defer func() { _ = resp.Body.Close() }()
-	if resp.StatusCode != http.StatusOK {
-		respBody, _ := io.ReadAll(resp.Body)
-		t.Fatalf("list status = %d, want 200: %s", resp.StatusCode, respBody)
-	}
-
-	var tokens []struct {
-		ID        string   `json:"id"`
-		Name      string   `json:"name"`
-		Scopes    []string `json:"scopes"`
-		CreatedAt string   `json:"createdAt"`
-		ExpiresAt string   `json:"expiresAt"`
-	}
-	if err := json.NewDecoder(resp.Body).Decode(&tokens); err != nil {
-		t.Fatalf("decode list response: %v", err)
-	}
-	if len(tokens) != 1 || tokens[0].ID != "grant-live" {
-		t.Fatalf("tokens = %+v, want only grant-live", tokens)
-	}
-	if tokens[0].Name != "grant-live" {
-		t.Fatalf("listed token name = %q, want grant id", tokens[0].Name)
-	}
-	if len(tokens[0].Scopes) != 1 || tokens[0].Scopes[0] != "testapp" {
-		t.Fatalf("listed scopes = %v, want [testapp]", tokens[0].Scopes)
-	}
-	expectedCreatedAt := time.Unix(now.Unix(), 0).UTC().Format(time.RFC3339)
-	if tokens[0].CreatedAt != expectedCreatedAt {
-		t.Fatalf("listed createdAt = %q, want %q", tokens[0].CreatedAt, expectedCreatedAt)
-	}
-	expectedExpiresAt := time.Unix(now.Add(time.Hour).Unix(), 0).UTC().Format(time.RFC3339)
-	if tokens[0].ExpiresAt != expectedExpiresAt {
-		t.Fatalf("listed expiresAt = %q, want %q", tokens[0].ExpiresAt, expectedExpiresAt)
-	}
-}
-
-func TestListAPITokensGetGrantFailureReturns500(t *testing.T) {
-	t.Parallel()
-
-	stub := newGrantTrackingAuthStub()
-	stub.ListGrantsFn = func(context.Context, *core.ListGrantsRequest) (*core.ListGrantsResponse, error) {
-		return &core.ListGrantsResponse{GrantIDs: []string{"grant-live"}}, nil
-	}
-	stub.GetGrantFn = func(context.Context, *core.GetGrantRequest) (*core.GetGrantResponse, error) {
-		return nil, fmt.Errorf("storage unavailable")
-	}
-
-	ts := newTestServer(t, func(cfg *server.Config) {
-		cfg.Auth = stub
-		cfg.Services = testutil.NewStubServices(t)
-	})
-	testutil.CloseOnCleanup(t, ts)
-
-	req, _ := http.NewRequest(http.MethodGet, ts.URL+"/api/v1/tokens", nil)
-	addGrantTestSessionCookie(req)
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		t.Fatalf("list request: %v", err)
-	}
-	defer func() { _ = resp.Body.Close() }()
-	if resp.StatusCode != http.StatusInternalServerError {
-		respBody, _ := io.ReadAll(resp.Body)
-		t.Fatalf("list status = %d, want 500: %s", resp.StatusCode, respBody)
-	}
-}
-
 func TestCreateAPITokenRejectsUnknownPermissionFields(t *testing.T) {
 	t.Parallel()
 
@@ -9741,75 +9549,6 @@ func TestCreateAPITokenRejectsUnknownPermissionFields(t *testing.T) {
 	}
 	if !strings.Contains(string(respBody), "invalid JSON body") {
 		t.Fatalf("create response = %s, want invalid JSON body error", respBody)
-	}
-}
-
-func TestListAPITokensListsOwnedUserRecords(t *testing.T) {
-	t.Parallel()
-
-	svc := testutil.NewStubServices(t)
-
-	ts := newTestServer(t, func(cfg *server.Config) {
-		configureGrantTestAuth(cfg)
-		cfg.Providers = grantTestProviders(t)
-		cfg.Services = svc
-	})
-	testutil.CloseOnCleanup(t, ts)
-
-	body := bytes.NewBufferString(`{"name":"owned-token","scopes":"testapp"}`)
-	req, _ := http.NewRequest(http.MethodPost, ts.URL+"/api/v1/tokens", body)
-	req.Header.Set("Content-Type", "application/json")
-	addGrantTestSessionCookie(req)
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		t.Fatalf("create request: %v", err)
-	}
-	defer func() { _ = resp.Body.Close() }()
-	if resp.StatusCode != http.StatusCreated {
-		respBody, _ := io.ReadAll(resp.Body)
-		t.Fatalf("create status = %d, want %d: %s", resp.StatusCode, http.StatusCreated, respBody)
-	}
-
-	var createResp struct {
-		ID string `json:"id"`
-	}
-	if err := json.NewDecoder(resp.Body).Decode(&createResp); err != nil {
-		t.Fatalf("decode create response: %v", err)
-	}
-
-	req, _ = http.NewRequest(http.MethodGet, ts.URL+"/api/v1/tokens", nil)
-	addGrantTestSessionCookie(req)
-	resp, err = http.DefaultClient.Do(req)
-	if err != nil {
-		t.Fatalf("list request: %v", err)
-	}
-	defer func() { _ = resp.Body.Close() }()
-	if resp.StatusCode != http.StatusOK {
-		respBody, _ := io.ReadAll(resp.Body)
-		t.Fatalf("list status = %d, want %d: %s", resp.StatusCode, http.StatusOK, respBody)
-	}
-
-	var tokens []struct {
-		ID        string   `json:"id"`
-		Name      string   `json:"name"`
-		Scopes    []string `json:"scopes"`
-		CreatedAt string   `json:"createdAt"`
-		ExpiresAt string   `json:"expiresAt"`
-	}
-	if err := json.NewDecoder(resp.Body).Decode(&tokens); err != nil {
-		t.Fatalf("decode list response: %v", err)
-	}
-	if len(tokens) != 1 || tokens[0].ID != createResp.ID {
-		t.Fatalf("tokens = %+v, want only %q", tokens, createResp.ID)
-	}
-	if len(tokens[0].Scopes) != 1 || tokens[0].Scopes[0] != "testapp" {
-		t.Fatalf("listed scopes = %v, want [testapp]", tokens[0].Scopes)
-	}
-	if _, err := time.Parse(time.RFC3339, tokens[0].CreatedAt); err != nil {
-		t.Fatalf("listed createdAt = %q, want RFC3339: %v", tokens[0].CreatedAt, err)
-	}
-	if _, err := time.Parse(time.RFC3339, tokens[0].ExpiresAt); err != nil {
-		t.Fatalf("listed expiresAt = %q, want RFC3339: %v", tokens[0].ExpiresAt, err)
 	}
 }
 
@@ -9884,121 +9623,6 @@ func TestRevokeAPIToken_WrongUser(t *testing.T) {
 	}
 }
 
-func TestCreateAPIToken_DefaultExpiry(t *testing.T) {
-	t.Parallel()
-
-	fixedNow := time.Date(2026, 1, 15, 12, 0, 0, 0, time.UTC)
-	svc := testutil.NewStubServices(t)
-	existing := seedUser(t, svc, "user@example.com")
-	var auditBuf bytes.Buffer
-	auditSink := invocation.NewSlogAuditSink(&auditBuf)
-	ts := newTestServer(t, func(cfg *server.Config) {
-		configureGrantTestAuthForUser(cfg, existing.ID)
-		cfg.Providers = grantTestProviders(t)
-		cfg.Now = func() time.Time { return fixedNow }
-		cfg.AuditSink = auditSink
-		cfg.Services = svc
-	})
-	testutil.CloseOnCleanup(t, ts)
-
-	body := bytes.NewBufferString(`{"name":"expiry-test","scopes":"testapp"}`)
-	req, _ := http.NewRequest(http.MethodPost, ts.URL+"/api/v1/tokens", body)
-	req.Header.Set("Content-Type", "application/json")
-	addGrantTestSessionCookie(req)
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		t.Fatalf("request: %v", err)
-	}
-	defer func() { _ = resp.Body.Close() }()
-
-	if resp.StatusCode != http.StatusCreated {
-		respBody, _ := io.ReadAll(resp.Body)
-		t.Fatalf("expected 201, got %d: %s", resp.StatusCode, respBody)
-	}
-
-	var result map[string]any
-	if err := json.NewDecoder(resp.Body).Decode(&result); err != nil {
-		t.Fatalf("decoding: %v", err)
-	}
-	tokenID, ok := result["id"].(string)
-	if !ok || tokenID == "" {
-		t.Fatalf("expected non-empty id in response, got %v", result["id"])
-	}
-	expiresAtRaw, ok := result["expiresAt"]
-	if !ok || expiresAtRaw == nil {
-		t.Fatal("expected expiresAt in response, got nil")
-	}
-	expiresAtStr, ok := expiresAtRaw.(string)
-	if !ok {
-		t.Fatalf("expected expiresAt to be a string, got %T", expiresAtRaw)
-	}
-	expiresAt, err := time.Parse(time.RFC3339, expiresAtStr)
-	if err != nil {
-		t.Fatalf("parsing expiresAt: %v", err)
-	}
-	expected := fixedNow.Add(30 * 24 * time.Hour).UTC().Truncate(time.Second)
-	if !expiresAt.Equal(expected) {
-		t.Fatalf("expected expiresAt %v, got %v", expected, expiresAt)
-	}
-
-	listReq, _ := http.NewRequest(http.MethodGet, ts.URL+"/api/v1/tokens", nil)
-	addGrantTestSessionCookie(listReq)
-	listResp, err := http.DefaultClient.Do(listReq)
-	if err != nil {
-		t.Fatalf("list grants: %v", err)
-	}
-	defer func() { _ = listResp.Body.Close() }()
-	if listResp.StatusCode != http.StatusOK {
-		respBody, _ := io.ReadAll(listResp.Body)
-		t.Fatalf("list status = %d, want 200: %s", listResp.StatusCode, respBody)
-	}
-	var grants []struct {
-		ID string `json:"id"`
-	}
-	if err := json.NewDecoder(listResp.Body).Decode(&grants); err != nil {
-		t.Fatalf("decode list response: %v", err)
-	}
-	if len(grants) != 1 {
-		t.Fatalf("expected 1 grant for canonical user, got %d", len(grants))
-	}
-	if grants[0].ID != tokenID {
-		t.Fatalf("expected stored grant ID %q, got %q", tokenID, grants[0].ID)
-	}
-
-	var auditRecord map[string]any
-	auditLines := bytes.Split(bytes.TrimSpace(auditBuf.Bytes()), []byte("\n"))
-	if len(auditLines) == 0 {
-		t.Fatal("expected grant.create audit record")
-	}
-	if err := json.Unmarshal(auditLines[0], &auditRecord); err != nil {
-		t.Fatalf("parsing audit record: %v\nraw: %s", err, auditBuf.String())
-	}
-	if auditRecord["operation"] != "grant.create" {
-		t.Fatalf("expected audit operation api_token.create, got %v", auditRecord["operation"])
-	}
-	if auditRecord["source"] != "http" {
-		t.Fatalf("expected audit source http, got %v", auditRecord["source"])
-	}
-	if subjectID, ok := auditRecord["subject_id"].(string); !ok || subjectID != principal.UserSubjectID(existing.ID) {
-		t.Fatalf("expected audit subject_id %q, got %v", principal.UserSubjectID(existing.ID), auditRecord["subject_id"])
-	}
-	if _, ok := auditRecord["user_id"]; ok {
-		t.Fatalf("expected emitted audit record to omit user_id, got %v", auditRecord["user_id"])
-	}
-	if auditRecord["allowed"] != true {
-		t.Fatalf("expected audit allowed=true, got %v", auditRecord["allowed"])
-	}
-	if auditRecord["target_kind"] != "api_token" {
-		t.Fatalf("expected audit target_kind api_token, got %v", auditRecord["target_kind"])
-	}
-	if auditRecord["target_id"] != tokenID {
-		t.Fatalf("expected audit target_id %q, got %v", tokenID, auditRecord["target_id"])
-	}
-	if auditRecord["target_name"] != "expiry-test" {
-		t.Fatalf("expected audit target_name expiry-test, got %v", auditRecord["target_name"])
-	}
-}
-
 func TestCreateAPIToken_AuditResolveUserFailure(t *testing.T) {
 	t.Parallel()
 
@@ -10044,237 +9668,6 @@ func TestCreateAPIToken_AuditResolveUserFailure(t *testing.T) {
 	}
 	if auditRecord["error"] != "failed to resolve user" {
 		t.Fatalf("expected audit error failed to resolve user, got %v", auditRecord["error"])
-	}
-}
-
-func TestRevokeAllAPITokens(t *testing.T) {
-	t.Parallel()
-
-	svc := testutil.NewStubServices(t)
-	u := seedUser(t, svc, "anonymous@gestalt")
-	var auditBuf bytes.Buffer
-	now := time.Now().UTC()
-	ts := newTestServer(t, func(cfg *server.Config) {
-		stub := configureGrantTestAuthForUser(cfg, u.ID)
-		for _, name := range []string{"tok-a", "tok-b", "tok-c"} {
-			stub.grants[name] = &core.GetGrantResponse{
-				CreatedAt: now.Unix(),
-				ExpiresAt: now.Add(24 * time.Hour).Unix(),
-			}
-		}
-		cfg.AuditSink = invocation.NewSlogAuditSink(&auditBuf)
-		cfg.Services = svc
-	})
-	testutil.CloseOnCleanup(t, ts)
-
-	req, _ := http.NewRequest(http.MethodDelete, ts.URL+"/api/v1/tokens", nil)
-	addGrantTestSessionCookie(req)
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		t.Fatalf("request: %v", err)
-	}
-	defer func() { _ = resp.Body.Close() }()
-
-	if resp.StatusCode != http.StatusOK {
-		t.Fatalf("expected 200, got %d", resp.StatusCode)
-	}
-
-	var result map[string]any
-	if err := json.NewDecoder(resp.Body).Decode(&result); err != nil {
-		t.Fatalf("decoding: %v", err)
-	}
-	if result["status"] != "revoked" {
-		t.Fatalf("expected status revoked, got %q", result["status"])
-	}
-	if count, ok := result["count"].(float64); !ok || count != 3 {
-		t.Fatalf("expected count 3, got %v", result["count"])
-	}
-
-	var auditRecord map[string]any
-	if err := json.Unmarshal(auditBuf.Bytes(), &auditRecord); err != nil {
-		t.Fatalf("parsing audit record: %v\nraw: %s", err, auditBuf.String())
-	}
-	if auditRecord["target_kind"] != "api_token_collection" {
-		t.Fatalf("expected audit target_kind api_token_collection, got %v", auditRecord["target_kind"])
-	}
-}
-
-func TestRevokeAllAPITokens_NoneExist(t *testing.T) {
-	t.Parallel()
-
-	svc := testutil.NewStubServices(t)
-	u := seedUser(t, svc, "anonymous@gestalt")
-	ts := newTestServer(t, func(cfg *server.Config) {
-		configureGrantTestAuthForUser(cfg, u.ID)
-		cfg.Services = svc
-	})
-	testutil.CloseOnCleanup(t, ts)
-
-	req, _ := http.NewRequest(http.MethodDelete, ts.URL+"/api/v1/tokens", nil)
-	addGrantTestSessionCookie(req)
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		t.Fatalf("request: %v", err)
-	}
-	defer func() { _ = resp.Body.Close() }()
-
-	if resp.StatusCode != http.StatusOK {
-		t.Fatalf("expected 200, got %d", resp.StatusCode)
-	}
-
-	var result map[string]any
-	if err := json.NewDecoder(resp.Body).Decode(&result); err != nil {
-		t.Fatalf("decoding: %v", err)
-	}
-	if count, ok := result["count"].(float64); !ok || count != 0 {
-		t.Fatalf("expected count 0, got %v", result["count"])
-	}
-}
-
-func TestRevokeAllAPITokens_ListGrantsNil(t *testing.T) {
-	t.Parallel()
-
-	svc := testutil.NewStubServices(t)
-	u := seedUser(t, svc, "anonymous@gestalt")
-	ts := newTestServer(t, func(cfg *server.Config) {
-		stub := configureGrantTestAuthForUser(cfg, u.ID)
-		stub.ListGrantsFn = func(context.Context, *core.ListGrantsRequest) (*core.ListGrantsResponse, error) {
-			return nil, nil
-		}
-		cfg.Services = svc
-	})
-	testutil.CloseOnCleanup(t, ts)
-
-	req, _ := http.NewRequest(http.MethodDelete, ts.URL+"/api/v1/tokens", nil)
-	addGrantTestSessionCookie(req)
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		t.Fatalf("request: %v", err)
-	}
-	defer func() { _ = resp.Body.Close() }()
-
-	if resp.StatusCode != http.StatusInternalServerError {
-		t.Fatalf("expected 500, got %d", resp.StatusCode)
-	}
-}
-
-func TestRevokeAllAPITokens_PartialFailure(t *testing.T) {
-	t.Parallel()
-
-	svc := testutil.NewStubServices(t)
-	u := seedUser(t, svc, "anonymous@gestalt")
-	now := time.Now().UTC()
-	ts := newTestServer(t, func(cfg *server.Config) {
-		stub := configureGrantTestAuthForUser(cfg, u.ID)
-		for _, name := range []string{"tok-a", "tok-b"} {
-			stub.grants[name] = &core.GetGrantResponse{
-				CreatedAt: now.Unix(),
-				ExpiresAt: now.Add(24 * time.Hour).Unix(),
-			}
-		}
-		stub.RevokeGrantFn = func(_ context.Context, req *core.RevokeGrantRequest) (*core.RevokeGrantResponse, error) {
-			if req == nil || req.GrantID == "" {
-				return nil, core.ErrNotFound
-			}
-			if req.GrantID == "tok-b" {
-				return nil, fmt.Errorf("failed to revoke grant")
-			}
-			stub.mu.Lock()
-			defer stub.mu.Unlock()
-			if _, ok := stub.grants[req.GrantID]; !ok {
-				return nil, core.ErrNotFound
-			}
-			stub.revoked[req.GrantID] = struct{}{}
-			return &core.RevokeGrantResponse{}, nil
-		}
-		cfg.Services = svc
-	})
-	testutil.CloseOnCleanup(t, ts)
-
-	req, _ := http.NewRequest(http.MethodDelete, ts.URL+"/api/v1/tokens", nil)
-	addGrantTestSessionCookie(req)
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		t.Fatalf("request: %v", err)
-	}
-	defer func() { _ = resp.Body.Close() }()
-
-	if resp.StatusCode != http.StatusInternalServerError {
-		t.Fatalf("expected 500, got %d", resp.StatusCode)
-	}
-
-	var result map[string]any
-	if err := json.NewDecoder(resp.Body).Decode(&result); err != nil {
-		t.Fatalf("decoding: %v", err)
-	}
-	if result["status"] != "partial" {
-		t.Fatalf("expected status partial, got %q", result["status"])
-	}
-	if count, ok := result["count"].(float64); !ok || count != 1 {
-		t.Fatalf("expected count 1, got %v", result["count"])
-	}
-	failed, ok := result["failed"].([]any)
-	if !ok || len(failed) != 1 {
-		t.Fatalf("expected one failed entry, got %v", result["failed"])
-	}
-	entry, ok := failed[0].(map[string]any)
-	if !ok {
-		t.Fatalf("expected failed entry object, got %T", failed[0])
-	}
-	if entry["id"] != "tok-b" {
-		t.Fatalf("expected failed id tok-b, got %v", entry["id"])
-	}
-}
-
-func TestRevokeAllAPITokens_ExcludesSessionGrants(t *testing.T) {
-	t.Parallel()
-
-	svc := testutil.NewStubServices(t)
-	u := seedUser(t, svc, "anonymous@gestalt")
-	now := time.Now().UTC()
-	var stub *grantTrackingAuthStub
-	ts := newTestServer(t, func(cfg *server.Config) {
-		stub = configureGrantTestAuthForUser(cfg, u.ID)
-		stub.grants["grant-stub"] = &core.GetGrantResponse{
-			CreatedAt: now.Unix(),
-			ExpiresAt: now.Add(24 * time.Hour).Unix(),
-		}
-		stub.grants["tok-a"] = &core.GetGrantResponse{
-			CreatedAt: now.Unix(),
-			ExpiresAt: now.Add(24 * time.Hour).Unix(),
-		}
-		cfg.Services = svc
-	})
-	testutil.CloseOnCleanup(t, ts)
-
-	req, _ := http.NewRequest(http.MethodDelete, ts.URL+"/api/v1/tokens", nil)
-	addGrantTestSessionCookie(req)
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		t.Fatalf("request: %v", err)
-	}
-	defer func() { _ = resp.Body.Close() }()
-
-	if resp.StatusCode != http.StatusOK {
-		t.Fatalf("expected 200, got %d", resp.StatusCode)
-	}
-
-	var result map[string]any
-	if err := json.NewDecoder(resp.Body).Decode(&result); err != nil {
-		t.Fatalf("decoding: %v", err)
-	}
-	if count, ok := result["count"].(float64); !ok || count != 1 {
-		t.Fatalf("expected count 1, got %v", result["count"])
-	}
-	stub.mu.Lock()
-	_, sessionRevoked := stub.revoked["grant-stub"]
-	_, apiRevoked := stub.revoked["tok-a"]
-	stub.mu.Unlock()
-	if sessionRevoked {
-		t.Fatal("expected session grant-stub to remain active")
-	}
-	if !apiRevoked {
-		t.Fatal("expected API token tok-a to be revoked")
 	}
 }
